telink: hal: drivers: ble: fix the operation of ble during pm
Using the manual mode of the system timer.
