+++ .gitignore
@@ -1,1 +1,1 @@
-**/blobs/liblt_9518_zephyr.a
+**/blobs
