+++ tlsr9/drivers/B92/stimer.h
@@ -49,7 +49,7 @@
  *                                           global macro                                                             *
  *********************************************************************************************************************/
 #ifndef SYS_TIMER_AUTO_MODE
-#define SYS_TIMER_AUTO_MODE     			1
+#define SYS_TIMER_AUTO_MODE     			0
 #endif
 /**********************************************************************************************************************
  *                                         global data type                                                           *
